fix: 설정 페이지 예산 토글 통합
