--- app/budgets/page.tsx
@@ -1,10 +1,0 @@
-import { MainLayout } from '@/components/layout/MainLayout';
-import { Budgets } from '@/components/budgets/Budgets';
-
-export default function BudgetsPage() {
-  return (
-    <MainLayout>
-      <Budgets />
-    </MainLayout>
-  );
-}
